Move tools into their own directory, following Go best practices. Also lets go-uic use gofmt from $PATH and moves dependencies for both go-uic and go-rcc into a new package named toolutil.
