+++ toolutil/util.go
@@ -1,4 +1,4 @@
-package main
+package toolutil
 
 import (
 	"os/exec"
@@ -6,7 +6,8 @@ import (
 	"github.com/kitech/qt.go/qtrt"
 )
 
-func runcmdout(name string, arg ...string) (string, error) {
+// RunCmdOut runs a comand and returns the commands output, or an error
+func RunCmdOut(name string, arg ...string) (string, error) {
 	cmd := exec.Command(name, arg...)
 	occ, err := cmd.Output() // Output runs the command and returns its standard output.
 	qtrt.ErrPrint(err, name, arg)
